script: Move some DOM interface conversion routines to script_bindings.
