+++ components/script/dom/bindings/conversions.rs
@@ -48,7 +48,6 @@ use crate::dom::bindings::error::{Error, Fallible};
 use crate::dom::bindings::reflector::DomObject;
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::trace::{JSTraceable, RootedTraceableBox};
-use crate::dom::bindings::utils::DOMClass;
 use crate::dom::filelist::FileList;
 use crate::dom::htmlcollection::HTMLCollection;
 use crate::dom::htmlformcontrolscollection::HTMLFormControlsCollection;
@@ -86,35 +85,6 @@ where
 
 pub(crate) use script_bindings::conversions::is_dom_proxy;
 
-/// Get a `*const libc::c_void` for the given DOM object, unless it is a DOM
-/// wrapper, and checking if the object is of the correct type.
-///
-/// Returns Err(()) if `obj` is a wrapper or if the object is not an object
-/// for a DOM object of the given type (as defined by the proto_id and proto_depth).
-#[inline]
-unsafe fn private_from_proto_check_static(
-    obj: *mut JSObject,
-    proto_check: fn(&'static DOMClass) -> bool,
-) -> Result<*const libc::c_void, ()> {
-    let dom_class = get_dom_class(obj).map_err(|_| ())?;
-    if proto_check(dom_class) {
-        trace!("good prototype");
-        Ok(private_from_object(obj))
-    } else {
-        trace!("bad prototype");
-        Err(())
-    }
-}
-
-/// Get a `*const T` for a DOM object accessible from a `JSObject`, where the DOM object
-/// is guaranteed not to be a wrapper.
-pub(crate) fn native_from_object_static<T>(obj: *mut JSObject) -> Result<*const T, ()>
-where
-    T: DomObject + IDLInterface,
-{
-    unsafe { private_from_proto_check_static(obj, T::derives).map(|ptr| ptr as *const T) }
-}
-
 /// Get a `DomRoot<T>` for the given DOM object, unwrapping any wrapper
 /// around it first, and checking if the object is of the correct type.
 ///
@@ -126,18 +96,6 @@ where
     T: DomObject + IDLInterface,
 {
     native_from_object_static(obj).map(|ptr| unsafe { DomRoot::from_ref(&*ptr) })
-}
-
-/// Get a `*const T` for a DOM object accessible from a `HandleValue`.
-/// Caller is responsible for throwing a JS exception if needed in case of error.
-pub(crate) fn native_from_handlevalue<T>(v: HandleValue, cx: *mut JSContext) -> Result<*const T, ()>
-where
-    T: DomObject + IDLInterface,
-{
-    if !v.get().is_object() {
-        return Err(());
-    }
-    unsafe { native_from_object(v.get().to_object(), cx) }
 }
 
 /// Get a `DomRoot<T>` for a DOM object accessible from a `HandleObject`.
